feat: CLIN-2661 Adjust Cypress tests
