--- cypress/e2e/Consultation/TableauVariants.cy.ts
@@ -1,245 +1,0 @@
-/// <reference types="Cypress" />
-import '../../support/commands';
-
-beforeEach(() => {
-  cy.login(Cypress.env('username_DG_CHUSJ_CUSM_CHUS'), Cypress.env('password'));
-});
-
-describe('Page des variants - Consultation du tableau', () => {
-  beforeEach(() => {
-    cy.visitVariantsPage('?sharedFilterId=ed4de9bb-016e-4869-ac9d-40b11ac3102a');
-    cy.showColumn('Tier', 0);
-    cy.showColumn('Max Fra.', 0);
-    cy.showColumn('Max Exo.', 0);
-    cy.showColumn('CADD', 0);
-    cy.showColumn('REVEL', 0);
-  });
-  
-  it('Vérifier les informations affichées', () => {
-    cy.validateTableDataRowKeyContent('4577893f4d3c2463e9fdef3419f7781d00fffdf3', 1, 'chrX:g.123403094G>A');
-    cy.validateTableDataRowKeyContent('4577893f4d3c2463e9fdef3419f7781d00fffdf3', 2, 'SNV');
-    cy.validateTableDataRowKeyClass('4577893f4d3c2463e9fdef3419f7781d00fffdf3', 3, 'anticon');
-    cy.validateTableDataRowKeyContent('4577893f4d3c2463e9fdef3419f7781d00fffdf3', 4, 'GRIA3');
-    cy.validateTableDataRowKeyAttr('4577893f4d3c2463e9fdef3419f7781d00fffdf3', 4, 'data-icon', 'plus');
-    cy.validateTableDataRowKeyClass('4577893f4d3c2463e9fdef3419f7781d00fffdf3', 5, 'ConsequencesCell_moderateImpact');
-    cy.validateTableDataRowKeyContent('4577893f4d3c2463e9fdef3419f7781d00fffdf3', 5, 'Missense p.Arg394Gln');
-    cy.get('tr[data-row-key="4577893f4d3c2463e9fdef3419f7781d00fffdf3"]').find('td').eq(6).find('path[d*="M16.7732"]').should('exist'); // C
-    cy.get('tr[data-row-key="4577893f4d3c2463e9fdef3419f7781d00fffdf3"]').find('td').eq(6).find('path[d*="M8.98279"]').should('exist'); // M
-    cy.get('tr[data-row-key="4577893f4d3c2463e9fdef3419f7781d00fffdf3"]').find('td').eq(6).find('path[d*="M10.9335"]').should('not.exist'); // P
-    cy.validateTableDataRowKeyContent('4577893f4d3c2463e9fdef3419f7781d00fffdf3', 7, 'XLR');
-    cy.validateTableDataRowKeyClass('4577893f4d3c2463e9fdef3419f7781d00fffdf3', 7, 'ant-tag-blue');
-    cy.validateTableDataRowKeyContent('4577893f4d3c2463e9fdef3419f7781d00fffdf3', 8, /^B$/);
-    cy.validateTableDataRowKeyContent('4577893f4d3c2463e9fdef3419f7781d00fffdf3', 8, /^LB$/);
-    cy.validateTableDataRowKeyClass('4577893f4d3c2463e9fdef3419f7781d00fffdf3', 8, 'ant-tag-green');
-    cy.validateTableDataRowKeyClass('4577893f4d3c2463e9fdef3419f7781d00fffdf3', 8, 'ant-tag-lime');
-    cy.validateTableDataRowKeyContent('4577893f4d3c2463e9fdef3419f7781d00fffdf3', 9, '9.01e-4');
-    cy.validateTableDataRowKeyClass('4577893f4d3c2463e9fdef3419f7781d00fffdf3', 9, 'GnomadCell_gnomadIndicator');
-    cy.validateTableDataRowKeyContent('4577893f4d3c2463e9fdef3419f7781d00fffdf3', 10, '101');
-    cy.validateTableDataRowKeyContent('4577893f4d3c2463e9fdef3419f7781d00fffdf3', 11, /^6$/);
-    cy.validateTableDataRowKeyContent('4577893f4d3c2463e9fdef3419f7781d00fffdf3', 11, /(4.\d{2}e-2)/);
-    cy.validateTableDataRowKeyContent('4577893f4d3c2463e9fdef3419f7781d00fffdf3', 12, /^3$/);
-    cy.validateTableDataRowKeyContent('4577893f4d3c2463e9fdef3419f7781d00fffdf3', 12, '(6.93e-5)');
-    cy.validateTableDataRowKeyContent('4577893f4d3c2463e9fdef3419f7781d00fffdf3', 13, '-');
-    cy.validateTableDataRowKeyContent('4577893f4d3c2463e9fdef3419f7781d00fffdf3', 14, '0.9436');
-    cy.validateTableDataRowKeyContent('4577893f4d3c2463e9fdef3419f7781d00fffdf3', 15, 'Other');
-    cy.validateTableDataRowKeyClass('4577893f4d3c2463e9fdef3419f7781d00fffdf3', 15, 'ant-tag-default');
-    cy.validateTableDataRowKeyContent('4577893f4d3c2463e9fdef3419f7781d00fffdf3', 16, '-');
-    cy.validateTableDataRowKeyContent('4577893f4d3c2463e9fdef3419f7781d00fffdf3', 17, '0.964');
-    cy.validateTableDataRowKeyContent('4577893f4d3c2463e9fdef3419f7781d00fffdf3', 18, '-');
-    cy.validateTableDataRowKeyContent('4577893f4d3c2463e9fdef3419f7781d00fffdf3', 19, 'VUS');
-    cy.validateTableDataRowKeyClass('4577893f4d3c2463e9fdef3419f7781d00fffdf3', 19, 'ant-tag-orange');
-    cy.validateTableDataRowKeyContent('4577893f4d3c2463e9fdef3419f7781d00fffdf3', 20, '-');
-    cy.validateTableDataRowKeyContent('4577893f4d3c2463e9fdef3419f7781d00fffdf3', 21, '2.56e+1');
-    cy.validateTableDataRowKeyContent('4577893f4d3c2463e9fdef3419f7781d00fffdf3', 22, '-');
-  });
- 
-  it('Valider l\'icône de sauvegarde des requêtes personnalisées', () => {
-    cy.get('[class*="QueryBar_selected"]').find('[class*="anticon-save"]').should('exist');
-  });
-
-  it('Valider les liens disponibles Lien Variant', () => {
-    cy.get('tr[data-row-key="4577893f4d3c2463e9fdef3419f7781d00fffdf3"]').contains('chrX:g.123403094G>A').invoke('removeAttr', 'target').click({force: true});
-    cy.get('[data-cy="Summary_Start"]').contains('123 403 094').should('exist');
-  });
- 
-  it('Valider les liens disponibles Lien dbSNP', () => {
-    cy.get('tr[data-row-key="4577893f4d3c2463e9fdef3419f7781d00fffdf3"]').find('td').eq(3).find('a[href]')
-      .should('have.attr', 'href', 'https://www.ncbi.nlm.nih.gov/snp/rs138817389');
-  });
- 
-  it('Valider les liens disponibles Lien Gène', () => {
-    cy.get('tr[data-row-key="4577893f4d3c2463e9fdef3419f7781d00fffdf3"]').find('td').eq(4).find('a[href]')
-      .should('have.attr', 'href', 'https://www.omim.org/entry/305915');
-  });
- 
-  it('Valider les liens disponibles Lien Gène Plus', () => {
-    cy.get('tr[data-row-key="4577893f4d3c2463e9fdef3419f7781d00fffdf3"]').find('td').eq(4).find('[data-icon="plus"]').click({force: true});
-    cy.validatePillSelectedQuery('Gène', ['GRIA3']);
-  });
- 
-  it('Valider les liens disponibles Lien OMIM', () => {
-    cy.get('tr[data-row-key="4577893f4d3c2463e9fdef3419f7781d00fffdf3"]').find('td').eq(7).find('a[href]')
-      .should('have.attr', 'href', 'https://www.omim.org/entry/305915');
-  });
- 
-  it('Valider les liens disponibles Lien ClinVar', () => {
-    cy.get('tr[data-row-key="4577893f4d3c2463e9fdef3419f7781d00fffdf3"]').find('td').eq(8).find('a[href]')
-      .should('have.attr', 'href', 'https://www.ncbi.nlm.nih.gov/clinvar/variation/198752');
-  });
- 
-  it('Valider les liens disponibles Lien RQDM', () => {
-    cy.get('tr[data-row-key="4577893f4d3c2463e9fdef3419f7781d00fffdf3"]').find('td').eq(11).find('a[href]').click({force: true});
-    cy.validateTableResultsCount('6 Résultats');
-  });
- 
-  it('Valider les liens disponibles Lien CMC', () => {
-    cy.get('tr[data-row-key="4577893f4d3c2463e9fdef3419f7781d00fffdf3"]').find('td').eq(12).find('a[href]')
-      .should('have.attr', 'href', 'https://cancer.sanger.ac.uk/cosmic/search?q=COSV52051792&genome=38#');
-  });
-});
-
-describe('Page des variants - Consultation du tableau', () => {
-  beforeEach(() => {
-    cy.visitVariantsPage('?sharedFilterId=0592969c-f83a-413a-b65d-578ab9d751fc');
-    cy.showColumn('Tier', 0);
-    cy.showColumn('Max Fra.', 0);
-    cy.showColumn('Max Exo.', 0);
-    cy.showColumn('CADD', 0);
-    cy.showColumn('REVEL', 0);
-  });
-  
-  it('Valider les fonctionnalités du tableau - Tri Variant', () => {
-    cy.waitWhileSpin(2000);
-
-    cy.sortTableAndIntercept('Variant', 1);
-    cy.validateTableFirstRow('chr10:g.100018446C>T', 1);
-    cy.sortTableAndIntercept('Variant', 1);
-    cy.validateTableFirstRow('chrY:g.9951709_9951710dup', 1);
-  });
-
-  it('Valider les fonctionnalités du tableau - Tri Type [CLIN-2149]', () => {
-    cy.waitWhileSpin(2000);
-
-    cy.sortTableAndIntercept('Type', 1);
-    cy.validateTableFirstRow('Ind', 2);
-    cy.sortTableAndIntercept('Type', 1);
-    cy.validateTableFirstRow('Sub', 2);
-  });
-
-  it('Valider les fonctionnalités du tableau - Tri Gène [CLIN-2149,CLIN-2287]', () => {
-    cy.waitWhileSpin(2000);
-
-    cy.sortTableAndIntercept('Gène', 1);
-    cy.validateTableFirstRow('-', 4);
-    cy.sortTableAndIntercept('Gène', 1);
-    cy.validateTableFirstRow('ZZZ3', 4);
-  });
-
-  it('Valider les fonctionnalités du tableau - Tri gnomAD', () => {
-    cy.waitWhileSpin(2000);
-
-    cy.sortTableAndIntercept(/^gnomAD $/, 1);
-    cy.validateTableFirstRow('-', 9);
-    cy.sortTableAndIntercept(/^gnomAD $/, 1);
-    cy.validateTableFirstRow('1.00e+0', 9);
-  });
-
-  it('Valider les fonctionnalités du tableau - Tri gnomAD ALT', () => {
-    cy.waitWhileSpin(2000);
-
-    cy.sortTableAndIntercept('gnomAD ALT', 1);
-    cy.validateTableFirstRow('-', 10);
-    cy.sortTableAndIntercept('gnomAD ALT', 1);
-    cy.validateTableFirstRow('152 312', 10);
-  });
-
-  it('Valider les fonctionnalités du tableau - Tri RQDM', () => {
-    cy.waitWhileSpin(2000);
-
-    cy.sortTableAndIntercept('RQDM', 1);
-    cy.validateTableFirstRow('0.00e+0', 11);
-    cy.sortTableAndIntercept('RQDM', 1);
-    cy.validateTableFirstRow(/(1\.00e\+0|9\.\d{2}e\-1)/, 11);
-  });
-
-  it('Valider les fonctionnalités du tableau - Tri CMC', () => {
-    cy.waitWhileSpin(2000);
-
-    cy.sortTableAndIntercept('CMC', 1);
-    cy.validateTableFirstRow('-', 12);
-    cy.sortTableAndIntercept('CMC', 1);
-    cy.validateTableFirstRow('419', 12);
-  });
-
-  it('Valider les fonctionnalités du tableau - Tri Hotspot', () => {
-    cy.waitWhileSpin(2000);
-
-    cy.clickAndIntercept('thead[class="ant-table-thead"] [data-icon="fire"]', 'POST', '**/graphql', 1);
-    cy.wait(1000);
-    cy.validateTableFirstRow('-', 13);
-    cy.clickAndIntercept('thead[class="ant-table-thead"] [data-icon="fire"]', 'POST', '**/graphql', 1);
-    cy.wait(1000);
-    cy.get('[class*="ant-table-row"]').eq(0).find('td').eq(13).find('[class*="hotspotFilled"]').should('exist');
-  });
-
-  it('Valider les fonctionnalités du tableau - Tri Exo. (var)', () => {
-    cy.waitWhileSpin(2000);
-
-    cy.sortTableAndIntercept('Exo. (var)', 1);
-    cy.validateTableFirstRow('-', 14);
-    cy.sortTableAndIntercept('Exo. (var)', 1);
-    cy.validateTableFirstRow('0.9992', 14);
-  });
-
-  it('Valider les fonctionnalités du tableau - Tri Tier', () => {
-    cy.waitWhileSpin(2000);
-
-    cy.sortTableAndIntercept('Tier', 1);
-    cy.validateTableFirstRow('-', 15);
-    cy.sortTableAndIntercept('Tier', 1);
-    cy.validateTableFirstRow('Other', 15);
-  });
-
-  it('Valider les fonctionnalités du tableau - Tri Max Exo.', () => {
-    cy.waitWhileSpin(2000);
-
-    cy.sortTableAndIntercept('Max Exo.', 1);
-    cy.validateTableFirstRow('-', 17);
-    cy.sortTableAndIntercept('Max Exo.', 1);
-    cy.validateTableFirstRow('0.964', 17);
-  });
-
-  it('Valider les fonctionnalités du tableau - Tri ACMG F.', () => {
-    cy.waitWhileSpin(2000);
-
-    cy.sortTableAndIntercept('ACMG F.', 1);
-    cy.validateTableFirstRow('-', 18);
-    cy.sortTableAndIntercept('ACMG F.', 1);
-    cy.validateTableFirstRow('VUS', 18);
-  });
-
-  it('Valider les fonctionnalités du tableau - Tri ACMG E.', () => {
-    cy.waitWhileSpin(2000);
-
-    cy.sortTableAndIntercept('ACMG E.', 1);
-    cy.validateTableFirstRow('-', 19);
-    cy.sortTableAndIntercept('ACMG E.', 1);
-    cy.validateTableFirstRow('VUS', 19);
-  });
-
-  it('Valider les fonctionnalités du tableau - Tri multiple', () => {
-    cy.waitWhileSpin(2000);
-
-    cy.sortTableAndIntercept('gnomAD', 1);
-    cy.sortTableAndWait('Variant');
-    cy.sortTableAndIntercept('Variant', 1);
-    cy.validateTableFirstRow('chrY:g.9906510G>C', 1);
-  });
-
-  it('Valider les fonctionnalités du tableau - Pagination', () => {
-    cy.waitWhileSpin(20000);
-
-    cy.validatePaging('', 0);
-  });
-});
-
