+++ cypress/e2e/Consultation/TableauVariantsPatientSomatic_2.cy.ts
@@ -1,0 +1,72 @@
+/// <reference types="Cypress" />
+import '../../support/commands';
+
+const presc_SOMATIC = JSON.parse(Cypress.env('presc_SOMATIC'));
+
+beforeEach(() => {
+  cy.login(Cypress.env('username_DG_CHUSJ_CUSM_CHUS'), Cypress.env('password'));
+  cy.visitVariantsPatientPage(presc_SOMATIC.patientProbId, presc_SOMATIC.prescriptionId, 3);
+
+  cy.showColumn('gnomAD ALT', 0);
+  cy.showColumn('QP', 0);
+  cy.showColumn(/^A$/, 0);
+  cy.showColumn('A+R', 0);
+  cy.showColumn('A/(A+R)', 0);
+  cy.showColumn('Filtre', 0);
+  cy.showColumn('CADD', 0);
+  cy.showColumn('REVEL', 0);
+});
+
+describe('Page des variants d\'un patient (somatic) - Consultation du tableau', () => { 
+  it('Valider les liens disponibles Lien UCSC', () => {
+    cy.get('tr[data-row-key="02fcc26c193333c0ed9f89fdfe6a3f79c5527af3"]').find('button[class*="ant-dropdown-trigger"]').click({force: true});
+    cy.get('[data-menu-id*="UCSC"]').find('a[href]')
+      .should('have.attr', 'href', 'https://genome.ucsc.edu/cgi-bin/hgTracks?db=hg38&lastVirtModeType=default&lastVirtModeExtraState=&virtModeType=default&virtMode=0&nonVirtPosition=&position=chr10%3A1096268-1096269');
+  });
+ 
+  it('Valider les liens disponibles Lien LitVar', () => {
+    cy.get('tr[data-row-key="02fcc26c193333c0ed9f89fdfe6a3f79c5527af3"]').find('button[class*="ant-dropdown-trigger"]').click({force: true});
+    cy.get('[data-menu-id*="LitVAR"]').find('a[href]')
+      .should('have.attr', 'href', 'https://www.ncbi.nlm.nih.gov/research/litvar2/docsum?text=rs10794716');
+  });
+ 
+  it('Valider les liens disponibles Lien Variant', () => {
+    cy.get('tr[data-row-key="02fcc26c193333c0ed9f89fdfe6a3f79c5527af3"]').contains('chr10:g.1096268T>C').invoke('removeAttr', 'target').click({force: true});
+    cy.get('[data-cy="Summary_Start"]').contains('1 096 268').should('exist');
+  });
+ 
+  it('Valider les liens disponibles Lien dbSNP', () => {
+    cy.get('tr[data-row-key="02fcc26c193333c0ed9f89fdfe6a3f79c5527af3"]').find('td').eq(5).find('a[href]')
+      .should('have.attr', 'href', 'https://www.ncbi.nlm.nih.gov/snp/rs10794716');
+  });
+ 
+  it('Valider les liens disponibles Lien Gène', () => {
+    cy.get('tr[data-row-key="02fcc26c193333c0ed9f89fdfe6a3f79c5527af3"]').find('td').eq(6).find('a[href]')
+      .should('have.attr', 'href', 'https://www.omim.org/entry/618586');
+  });
+ 
+  it('Valider les liens disponibles Lien Gène Plus', () => {
+    cy.get('tr[data-row-key="02fcc26c193333c0ed9f89fdfe6a3f79c5527af3"]').find('td').eq(6).find('[data-icon="plus"]').click({force: true});
+    cy.validatePillSelectedQuery('Gène', ['WDR37']);
+  });
+ 
+  it('Valider les liens disponibles Lien OMIM', () => {
+    cy.get('tr[data-row-key="02fcc26c193333c0ed9f89fdfe6a3f79c5527af3"]').find('td').eq(9).find('a[href]')
+      .should('have.attr', 'href', 'https://www.omim.org/entry/618586');
+  });
+ 
+  it('Valider les liens disponibles Lien ClinVar', () => {
+    cy.get('tr[data-row-key="02fcc26c193333c0ed9f89fdfe6a3f79c5527af3"]').find('td').eq(10).find('a[href]')
+      .should('have.attr', 'href', 'https://www.ncbi.nlm.nih.gov/clinvar/variation/1185321');
+  });
+ 
+  it('Valider les liens disponibles Lien CMC', () => {
+    cy.get('tr[data-row-key="bdc7b7f2fba4aef570b1ac84217fe870f14261db"]').find('td').eq(13).find('a[href]')
+      .should('have.attr', 'href', 'https://cancer.sanger.ac.uk/cosmic/search?q=COSV53313027&genome=38#');
+  });
+ 
+  it('Valider les liens disponibles Lien RQDM', () => {
+    cy.get('tr[data-row-key="02fcc26c193333c0ed9f89fdfe6a3f79c5527af3"]').find('td').eq(16).find('a[href]').click({force: true});
+    cy.validateTableResultsCount(/^13\d{1}$/);
+  });
+});
