+++ cypress/e2e/Consultation/TableauVariantsPatient_3.cy.ts
@@ -1,0 +1,227 @@
+/// <reference types="Cypress" />
+import '../../support/commands';
+
+const epCHUSJ_ldmCHUSJ = JSON.parse(Cypress.env('presc_EP_CHUSJ_LDM_CHUSJ'));
+
+beforeEach(() => {
+  cy.login(Cypress.env('username_DG_CHUSJ_CUSM_CHUS'), Cypress.env('password'));
+  cy.visitVariantsPatientPage(epCHUSJ_ldmCHUSJ.patientProbId, epCHUSJ_ldmCHUSJ.prescriptionId, 3);
+
+  cy.showColumn('M : P', 0);
+  cy.showColumn('HC', 0);
+  cy.showColumn('HCP', 0);
+  cy.showColumn('Trans.', 0);
+  cy.showColumn('QP', 0);
+  cy.showColumn('OP', 0);
+  cy.showColumn(/^A$/, 0);
+  cy.showColumn('A+R', 0);
+  cy.showColumn('A/(A+R)', 0);
+  cy.showColumn('Filtre', 0);
+  cy.showColumn('Crit. Exo.', 0);
+  cy.showColumn(/^CMC$/, 0);
+  cy.showColumn('Tier', 0);
+  cy.showColumn('CADD', 0);
+  cy.showColumn('REVEL', 0);
+  cy.showColumn('Crit. Fra.', 0);
+});
+
+describe('Page des variants d\'un patient - Consultation du tableau', () => {  
+  it('Valider les fonctionnalités du tableau - Tri Variant', () => {
+    cy.get('[id="query-builder-header-tools"]').find('[data-icon="plus"]').click({force: true});
+    cy.waitWhileSpin(2000);
+
+    cy.sortTableAndIntercept('Variant', 1);
+    cy.validateTableFirstRow('chr10:g.100049106G>A', 2);
+    cy.sortTableAndIntercept('Variant', 1);
+    cy.validateTableFirstRow('chrY:g.9951709_9951710del', 2);
+  });
+
+  it('Valider les fonctionnalités du tableau - Tri Type [CLIN-2149]', () => {
+    cy.get('[id="query-builder-header-tools"]').find('[data-icon="plus"]').click({force: true});
+    cy.waitWhileSpin(2000);
+
+    cy.sortTableAndIntercept('Type', 1);
+    cy.validateTableFirstRow('Ins', 3);
+    cy.sortTableAndIntercept('Type', 1);
+    cy.validateTableFirstRow('SNV', 3);
+  });
+
+  it('Valider les fonctionnalités du tableau - Tri Gène', () => {
+    cy.get('[id="query-builder-header-tools"]').find('[data-icon="plus"]').click({force: true});
+    cy.waitWhileSpin(2000);
+
+    cy.sortTableAndIntercept('Gène', 1);
+    cy.validateTableFirstRow('-', 5);
+    cy.sortTableAndIntercept('Gène', 1);
+    cy.validateTableFirstRow('ZZZ3', 5);
+  });
+
+  it('Valider les fonctionnalités du tableau - Tri Fra.', () => {
+    cy.get('[id="query-builder-header-tools"]').find('[data-icon="plus"]').click({force: true});
+    cy.waitWhileSpin(2000);
+
+    cy.sortTableAndIntercept('Fra.', 1);
+    cy.validateTableFirstRow('-', 10);
+    cy.sortTableAndIntercept('Fra.', 1);
+    cy.validateTableFirstRow('-', 10);
+  });
+
+  it('Valider les fonctionnalités du tableau - Tri Exo.', () => {
+    cy.get('[id="query-builder-header-tools"]').find('[data-icon="plus"]').click({force: true});
+    cy.waitWhileSpin(2000);
+
+    cy.sortTableAndIntercept('Exo.', 1);
+    cy.validateTableFirstRow('-', 11);
+    cy.sortTableAndIntercept('Exo.', 1);
+    cy.validateTableFirstRow('0.964', 11);
+  });
+
+  it('Valider les fonctionnalités du tableau - Tri ACMG F.', () => {
+    cy.get('[id="query-builder-header-tools"]').find('[data-icon="plus"]').click({force: true});
+    cy.waitWhileSpin(2000);
+
+    cy.sortTableAndIntercept('ACMG F.', 1);
+    cy.validateTableFirstRow('-', 12);
+    cy.sortTableAndIntercept('ACMG F.', 1);
+    cy.validateTableFirstRow('VUS', 12);
+  });
+
+  it('Valider les fonctionnalités du tableau - Tri ACMG E.', () => {
+    cy.get('[id="query-builder-header-tools"]').find('[data-icon="plus"]').click({force: true});
+    cy.waitWhileSpin(2000);
+
+    cy.sortTableAndIntercept('ACMG E.', 1);
+    cy.validateTableFirstRow('-', 13);
+    cy.sortTableAndIntercept('ACMG E.', 1);
+    cy.validateTableFirstRow('VUS', 13);
+  });
+
+  it('Valider les fonctionnalités du tableau - Tri gnomAD', () => {
+    cy.get('[id="query-builder-header-tools"]').find('[data-icon="plus"]').click({force: true});
+    cy.waitWhileSpin(2000);
+
+    cy.sortTableAndIntercept(/^gnomAD $/, 1);
+    cy.validateTableFirstRow('-', 14);
+    cy.sortTableAndIntercept(/^gnomAD $/, 1);
+    cy.validateTableFirstRow('1.00e+0', 14);
+  });
+
+  it('Valider les fonctionnalités du tableau - Tri gnomAD ALT', () => {
+    cy.get('[id="query-builder-header-tools"]').find('[data-icon="plus"]').click({force: true});
+    cy.waitWhileSpin(2000);
+
+    cy.sortTableAndIntercept('gnomAD ALT', 1);
+    cy.validateTableFirstRow('-', 15);
+    cy.sortTableAndIntercept('gnomAD ALT', 1);
+    cy.validateTableFirstRow('152 312', 15);
+  });
+
+  it('Valider les fonctionnalités du tableau - Tri RQDM', () => {
+    cy.get('[id="query-builder-header-tools"]').find('[data-icon="plus"]').click({force: true});
+    cy.waitWhileSpin(2000);
+
+    cy.sortTableAndIntercept('RQDM', 1);
+    cy.validateTableFirstRow('0.00e+0', 16);
+    cy.sortTableAndIntercept('RQDM', 1);
+    cy.validateTableFirstRow(/(1\.00e\+0|9\.\d{2}e\-1)/, 16);
+  });
+
+  it('Valider les fonctionnalités du tableau - Tri QG', () => {
+    cy.sortTableAndIntercept('QG', 1);
+    cy.validateTableFirstRow(/^1$/, 17);
+    cy.sortTableAndIntercept('QG', 1);
+    cy.validateTableFirstRow('447', 17);
+  });
+
+  it('Valider les fonctionnalités du tableau - Tri Zyg.', () => {
+    cy.sortTableAndIntercept('Zyg.', 1);
+    cy.validateTableFirstRow(/^1$/, 18);
+    cy.sortTableAndIntercept('Zyg.', 1);
+    cy.validateTableFirstRow('1/1', 18);
+  });
+
+  it('Valider les fonctionnalités du tableau - Tri Trans.', () => {
+    cy.sortTableAndIntercept('Trans.', 1);
+    cy.validateTableFirstRow('-', 22);
+    cy.sortTableAndIntercept('Trans.', 1);
+    cy.validateTableFirstRow('XLR DNV', 22);
+  });
+
+  it('Valider les fonctionnalités du tableau - Tri QP', () => {
+    cy.sortTableAndIntercept('QP', 1);
+    cy.validateTableFirstRow(/^0$/, 23);
+    cy.sortTableAndIntercept('QP', 1);
+    cy.validateTableFirstRow('34.99', 23);
+  });
+
+  it('Valider les fonctionnalités du tableau - Tri OP', () => {
+    cy.sortTableAndIntercept('OP', 1);
+    cy.validateTableFirstRow('-', 24);
+    cy.sortTableAndIntercept('OP', 1);
+    cy.validateTableFirstRow('UNK', 24);
+  });
+
+  it('Valider les fonctionnalités du tableau - Tri A', () => {
+    cy.sortTableAndIntercept(/^A$/, 1);
+    cy.validateTableFirstRow(/^3$/, 25);
+    cy.sortTableAndIntercept(/^A$/, 1);
+    cy.validateTableFirstRow('458', 25);
+  });
+
+  it('Valider les fonctionnalités du tableau - Tri A+R', () => {
+    cy.sortTableAndIntercept(/^A\+R$/, 1);
+    cy.validateTableFirstRow(/^3$/, 26);
+    cy.sortTableAndIntercept(/^A\+R$/, 1);
+    cy.validateTableFirstRow('631', 26);
+  });
+
+  it('Valider les fonctionnalités du tableau - Tri A/(A+R)', () => {
+    cy.sortTableAndIntercept('A/(A+R)', 1);
+    cy.validateTableFirstRow('0.05', 27);
+    cy.sortTableAndIntercept('A/(A+R)', 1);
+    cy.validateTableFirstRow('1.00', 27);
+  });
+
+  it('Valider les fonctionnalités du tableau - Tri Filtre', () => {
+    cy.sortTableAndIntercept('Filtre', 1);
+    cy.validateTableFirstRow('DRAGENIndelHardQUAL', 28);
+    cy.sortTableAndIntercept('Filtre', 1);
+    cy.validateTableFirstRow('PASS', 28);
+  });
+
+  it('Valider les fonctionnalités du tableau - Tri CMC', () => {
+    cy.get('[id="query-builder-header-tools"]').find('[data-icon="plus"]').click({force: true});
+    cy.waitWhileSpin(2000);
+
+    cy.sortTableAndIntercept('CMC', 1);
+    cy.validateTableFirstRow('-', 30);
+    cy.sortTableAndIntercept('CMC', 1);
+    cy.validateTableFirstRow('419', 30);
+  });
+
+  it('Valider les fonctionnalités du tableau - Tri Tier', () => {
+    cy.get('[id="query-builder-header-tools"]').find('[data-icon="plus"]').click({force: true});
+    cy.waitWhileSpin(2000);
+
+    cy.sortTableAndIntercept('Tier', 1);
+    cy.validateTableFirstRow('-', 31);
+    cy.sortTableAndIntercept('Tier', 1);
+    cy.validateTableFirstRow('Other', 31);
+  });
+
+  it('Valider les fonctionnalités du tableau - Tri multiple', () => {
+    cy.get('[id="query-builder-header-tools"]').find('[data-icon="plus"]').click({force: true});
+    cy.waitWhileSpin(2000);
+
+    cy.sortTableAndIntercept('gnomAD', 1);
+    cy.sortTableAndIntercept('gnomAD', 1);
+    cy.sortTableAndIntercept('Gène', 1);
+    cy.sortTableAndIntercept('Gène', 1);
+    cy.validateTableFirstRow(/(ZNHIT1|ZNF875)/, 5);
+  });
+
+  it('Valider les fonctionnalités du tableau - Pagination', () => {
+    cy.validatePaging('', 1);
+  });
+});
+
