+++ cypress/e2e/Consultation/TableauVariants_2.cy.ts
@@ -1,0 +1,58 @@
+/// <reference types="Cypress" />
+import '../../support/commands';
+
+beforeEach(() => {
+  cy.login(Cypress.env('username_DG_CHUSJ_CUSM_CHUS'), Cypress.env('password'));
+});
+
+describe('Page des variants - Consultation du tableau', () => {
+  beforeEach(() => {
+    cy.visitVariantsPage('?sharedFilterId=ed4de9bb-016e-4869-ac9d-40b11ac3102a');
+    cy.showColumn('Tier', 0);
+    cy.showColumn('Max Fra.', 0);
+    cy.showColumn('Max Exo.', 0);
+    cy.showColumn('CADD', 0);
+    cy.showColumn('REVEL', 0);
+  });
+
+  it('Valider les liens disponibles Lien Variant', () => {
+    cy.get('tr[data-row-key="4577893f4d3c2463e9fdef3419f7781d00fffdf3"]').contains('chrX:g.123403094G>A').invoke('removeAttr', 'target').click({force: true});
+    cy.get('[data-cy="Summary_Start"]').contains('123 403 094').should('exist');
+  });
+ 
+  it('Valider les liens disponibles Lien dbSNP', () => {
+    cy.get('tr[data-row-key="4577893f4d3c2463e9fdef3419f7781d00fffdf3"]').find('td').eq(3).find('a[href]')
+      .should('have.attr', 'href', 'https://www.ncbi.nlm.nih.gov/snp/rs138817389');
+  });
+ 
+  it('Valider les liens disponibles Lien Gène', () => {
+    cy.get('tr[data-row-key="4577893f4d3c2463e9fdef3419f7781d00fffdf3"]').find('td').eq(4).find('a[href]')
+      .should('have.attr', 'href', 'https://www.omim.org/entry/305915');
+  });
+ 
+  it('Valider les liens disponibles Lien Gène Plus', () => {
+    cy.get('tr[data-row-key="4577893f4d3c2463e9fdef3419f7781d00fffdf3"]').find('td').eq(4).find('[data-icon="plus"]').click({force: true});
+    cy.validatePillSelectedQuery('Gène', ['GRIA3']);
+  });
+ 
+  it('Valider les liens disponibles Lien OMIM', () => {
+    cy.get('tr[data-row-key="4577893f4d3c2463e9fdef3419f7781d00fffdf3"]').find('td').eq(7).find('a[href]')
+      .should('have.attr', 'href', 'https://www.omim.org/entry/305915');
+  });
+ 
+  it('Valider les liens disponibles Lien ClinVar', () => {
+    cy.get('tr[data-row-key="4577893f4d3c2463e9fdef3419f7781d00fffdf3"]').find('td').eq(8).find('a[href]')
+      .should('have.attr', 'href', 'https://www.ncbi.nlm.nih.gov/clinvar/variation/198752');
+  });
+ 
+  it('Valider les liens disponibles Lien RQDM', () => {
+    cy.get('tr[data-row-key="4577893f4d3c2463e9fdef3419f7781d00fffdf3"]').find('td').eq(11).find('a[href]').click({force: true});
+    cy.validateTableResultsCount('6 Résultats');
+  });
+ 
+  it('Valider les liens disponibles Lien CMC', () => {
+    cy.get('tr[data-row-key="4577893f4d3c2463e9fdef3419f7781d00fffdf3"]').find('td').eq(12).find('a[href]')
+      .should('have.attr', 'href', 'https://cancer.sanger.ac.uk/cosmic/search?q=COSV52051792&genome=38#');
+  });
+});
+
