+++ cypress/e2e/Facettes/PageCNVsPatient.cy.ts
@@ -17,6 +17,61 @@ describe('Page des CNVs d\'un patient - Filtrer avec les facettes', () => {
     cy.validateExpandCollapse('Panel RQDM');
   });
 
+  it('Recherche par gène - CFHR1', () => {
+    cy.get(`[data-cy="SidebarMenuItem_Gène"]`).click({force: true});
+    cy.get('[class*="SearchLabel_title"]').contains('Recherche par gène').should('exist'); //data-cy="SearchLabel_Title"
+
+    cy.get('[class*="SearchLabel_tooltipIcon"]').trigger('mouseover', {eventConstructor: 'MouseEvent', force: true}); //data-cy="SearchLabel_InfoCircleOutlined"
+    cy.get('body').contains('Recherche par symbole, alias et Ensembl ID').should('exist');
+
+    cy.intercept('GET', '**/CFHR1').as('getRouteMatcher');
+    cy.get('[class*="SearchAutocomplete_search"]').eq(0).find('input').type('CFHR1', {force: true}); //data-cy="SearchAutocomplete_Select"
+    cy.wait('@getRouteMatcher', {timeout: 60*1000});
+
+    cy.get('[class*="ant-select-dropdown"]').contains('CFHR1').should('exist'); //data-cy="Search_Dropdown"
+    cy.get('[class*="ant-select-dropdown"]').find('div[class*="ant-select-item"]').eq(0).click({force: true}); //data-cy="Search_Dropdown"
+
+    cy.get('[class*="ant-tag"]').contains('CFHR1').should('exist'); //data-cy="Tag_CFHR1"
+    cy.get('[class*="QueryBar_selected"]').find('[class*="QueryPill_field"]').contains('Gène').should('exist');
+    cy.get('[class*="QueryBar_selected"]').find('[class*="QueryValues_value"]').contains('CFHR1').should('exist');
+    cy.validateTableResultsCount(/^1$/);
+
+    cy.get('[data-icon="close-circle"]').click({force: true});
+    cy.get('[class*="ant-select-show-search"] [class="ant-tag"]').should('not.exist'); //data-cy="Tag_CFHR1"
+  });
+
+  it('Recherche par gène - FHR1', () => {
+    cy.get(`[data-cy="SidebarMenuItem_Gène"]`).click({force: true});
+
+    cy.intercept('GET', '**/FHR1').as('getRouteMatcher');
+    cy.get('[class*="SearchAutocomplete_search"]').eq(0).find('input').type('FHR1', {force: true}); //data-cy="SearchAutocomplete_Select"
+    cy.wait('@getRouteMatcher', {timeout: 60*1000});
+
+    cy.get('[class*="ant-select-dropdown"]').contains('CFHR1').should('exist'); //data-cy="Search_Dropdown"
+    cy.get('[class*="ant-select-dropdown"]').find('div[class*="ant-select-item"]').eq(0).click({force: true}); //data-cy="Search_Dropdown"
+
+    cy.get('[class*="ant-tag"]').contains('CFHR1').should('exist'); //data-cy="Tag_CFHR1"
+    cy.get('[class*="QueryBar_selected"]').find('[class*="QueryPill_field"]').contains('Gène').should('exist');
+    cy.get('[class*="QueryBar_selected"]').find('[class*="QueryValues_value"]').contains('CFHR1').should('exist');
+    cy.validateTableResultsCount(/^1$/);
+  });
+
+  it('Recherche par gène - ENSG00000244414', () => {
+    cy.get(`[data-cy="SidebarMenuItem_Gène"]`).click({force: true});
+
+    cy.intercept('GET', '**/ENSG00000244414').as('getRouteMatcher');
+    cy.get('[class*="SearchAutocomplete_search"]').eq(0).find('input').type('ENSG00000244414', {force: true}); //data-cy="SearchAutocomplete_Select"
+    cy.wait('@getRouteMatcher', {timeout: 60*1000});
+
+    cy.get('[class*="ant-select-dropdown"]').contains('CFHR1').should('exist'); //data-cy="Search_Dropdown"
+    cy.get('[class*="ant-select-dropdown"]').find('div[class*="ant-select-item"]').eq(0).click({force: true}); //data-cy="Search_Dropdown"
+
+    cy.get('[class*="ant-tag"]').contains('CFHR1').should('exist'); //data-cy="Tag_CFHR1"
+    cy.get('[class*="QueryBar_selected"]').find('[class*="QueryPill_field"]').contains('Gène').should('exist');
+    cy.get('[class*="QueryBar_selected"]').find('[class*="QueryValues_value"]').contains('CFHR1').should('exist');
+    cy.validateTableResultsCount(/^1$/);
+  });
+
   it('Panel RQDM - Panel RQDM', () => {
     cy.validateFacetFilter('Panel RQDM', 'Panel RQDM', 'POLYM', 'POLYM', /^26$/);
     cy.validateFacetRank(0, 'Panel RQDM');
