+++ cypress/e2e/Facettes/PageVariantsPatient_1.cy.ts
@@ -1,0 +1,24 @@
+/// <reference types="Cypress" />
+import '../../support/commands';
+
+const epCHUSJ_ldmCHUSJ = JSON.parse(Cypress.env('presc_EP_CHUSJ_LDM_CHUSJ'));
+
+beforeEach(() => {
+  cy.login(Cypress.env('username_DG_CHUSJ_CUSM_CHUS'), Cypress.env('password'));
+});
+
+describe('Page des variants d\'un patient - Filtrer avec les facettes', () => {
+
+  beforeEach(() => {
+    cy.visitVariantsPatientPage(epCHUSJ_ldmCHUSJ.patientProbId, epCHUSJ_ldmCHUSJ.prescriptionId, 3);
+  });
+
+  it('Panel RQDM - Expand all/Collapse all', () => {
+    cy.validateExpandCollapse('Panel RQDM');
+  });
+
+  it('Panel RQDM - RQDM', () => {
+    cy.validateFacetFilter('Panel RQDM', 'RQDM', 'POLYM', 'POLYM', /^24 975$/);
+    cy.validateFacetRank(0, 'RQDM');
+  });
+});
